+++ packages/media/media-ui/src/LinkView/index.tsx
@@ -1,14 +1,12 @@
 import * as React from 'react';
 import { Link } from './styled';
 
-export interface InlineCardLinkViewProps {
+export interface CardLinkViewProps {
   text: string;
   onClick?: () => void;
 }
 
-export class InlineCardLinkView extends React.Component<
-  InlineCardLinkViewProps
-> {
+export class CardLinkView extends React.Component<CardLinkViewProps> {
   handleClick = (event: React.MouseEvent<HTMLAnchorElement>) => {
     const { onClick } = this.props;
     if (onClick) {