+++ packages/media/media-avatar-picker/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-avatar-picker
 
+## 12.0.3
+- [patch] [5a6de24](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/5a6de24):
+
+  - translate component properties in media components
+
 ## 12.0.2
 - Updated dependencies [ab9b69c](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ab9b69c):
   - @atlaskit/docs@5.2.2